release v0.1.1

+++ test_files/test_8_fail_import_in_album.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	struct timeval nstime;
-	char tmp[300] = {0};
+	char tmp[4096] = {0};
 	char lib_name[50];
 
 	gettimeofday(&nstime, NULL);
@@ -37,11 +37,26 @@ int main(void)
 		printf("no error executing ospl_import_photo_in_album\n");
 		return 0;
 	}
+	if (ospl_album_add_photo(lib_name, 1, 1) >= 0) 
+	{
+		printf("no error executing ospl_album_add_photo\n");
+		return 0;
+	}
+	if (ospl_album_add_photo(lib_name, 1, 10) >= 0) 
+	{
+		printf("no error executing ospl_album_add_photo\n");
+		return 0;
+	}
+	if (ospl_import_photo_in_album(lib_name, "../test_files/ressources/1b3BAEr6IJk.jpg", 50) >= 0)
+	{
+		printf("no error executing ospl_import_photo_in_album\n");
+		return 0;
+	}
 	sprintf(tmp, "rm -rf %s", lib_name);
 	if (system(tmp))
 	{
 		printf("executing the system command failed\n");
-		return 1;
+		return 0;
 	}
 	return 1;
 }
